drop Go 1.7 support
